Remove unnecessary casting
git-svn-id: http://svn.lynbrookrobotics.com/code15@209 9f8849d2-cf13-4b18-a094-84ffabbe6057

+++ Robot/LRT15/src/com/team846/frc2015/sensors/DriveEncoders.java
@@ -135,7 +135,7 @@ public class DriveEncoders implements Configurable
 	public double GetWheelDist(Side side)
 	{
 		LRTCANEncoder e = encoders[side.ordinal()];
-		double dist = (double) ((e.get()) / PULSES_PER_REVOLUTION
+		double dist = ((e.get()) / PULSES_PER_REVOLUTION
 				* GEAR_RATIO * WHEEL_DIAMETER * Math.PI); // pulses / ( pulses / encoder revolution ) * encoder to wheel gear ratio * distance / wheel revolution = inch distance
 		return dist;
 	}
